python: remove tests and sqlite3-tests modules
These modules are most certainly useless for a production setup
(from python docs: 'The test package is meant for internal use
by Python only') and take a significant amount of disk space
(roughly 20M for 'test').

If present in manifests, they are shipped in 'python-modules'. This
commit keeps them avalaible in the 'python-misc' package.

Signed-off-by: Laurent Bonnans <laurent.bonnans@hikob.com>